pyproject.toml
[tool.poetry]
name = "agentq"
version = "0.1.0"
description = ""
authors = ["nischalj10 <nischalj10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
litellm = "^1.42.9"
pydantic = "^2.8.2"
pytest-playwright = "^0.5.1"
pdfplumber = "0.11.2"
typing-extensions = "^4.12.2"
ruff = "^0.5.6"
playwright-stealth = "^1.0.6"
setuptools = "^72.1.0"
openai = "^1.40.1"
boto3 = "^1.34.157"
python-json-logger = "^2.0.7"
aiohttp = "^3.10.2"
colorama = "^0.4.6"
tiktoken = "^0.7.0"
agentops = "^0.3.10"
termcolor = "^2.4.0"
tabulate = "^0.9.0"
nltk = "^3.9.1"
langsmith = "^0.1.104"
instructor = "^1.4.0"
flask = "^3.0.3"
numpy = "^2.1.0"
python-dotenv = "^1.0.1"
smolagents = {version = "1.18.0", extras = ["toolkit"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
